# silvera/sciLibrary/output/SciPaper/openapi.json
{"openapi": "3.0.0", "info": {"version": "0.0.1b", "title": "SciPaper", "description": ""}, "servers": [{"url": "http://localhost:9191/scipaper"}], "paths": {"/scipaper": {"post": {"description": "Creates a new scipaper.", "operationId": "createSciPaper", "requestBody": {"description": "SciPaper to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}}, "/scipaper/id": {"put": {"description": "Update SciPaper object.", "operationId": "updateSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}, "get": {"description": "Read SciPaper object.", "operationId": "readSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}, "delete": {"description": "Delete SciPaper object.", "operationId": "deleteSciPaper", "parameters": [{"name": "id", "in": "path", "description": "ID of scipaper to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SciPaperBody"}, "responses": {"200": {"$ref": "#components/responses/SciPaperOK"}}}}, "/section": {"post": {"description": "Creates a new section.", "operationId": "createSection", "requestBody": {"description": "Section to add.", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}, "responses": {"200": {"$ref": "#components/responses/SectionOK"}}}}, "/section/id": {"put": {"description": "Update Section object.", "operationId": "updateSection", "parameters": [{"name": "id", "in": "path", "description": "ID of section to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SectionBody"}, "responses": {"200": {"$ref": "#components/responses/SectionOK"}}}, "get": {"description": "Read Section object.", "operationId": "readSection", "parameters": [{"name": "id", "in": "path", "description": "ID of section to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SectionBody"}, "responses": {"200": {"$ref": "#components/responses/SectionOK"}}}, "delete": {"description": "Delete Section object.", "operationId": "deleteSection", "parameters": [{"name": "id", "in": "path", "description": "ID of section to update", "required": true, "schema": {"type": "string"}}], "requestBody": {"$ref": "#components/requestBodies/SectionBody"}, "responses": {"200": {"$ref": "#components/responses/SectionOK"}}}}, "//{username}": {"post": {"description": "", "operationId": "uploadSciPaper", "parameters": [{"name": "username", "in": "path", "required": true, "schema": {"type": "string"}}], "requestBody": {"description": "JSON representation object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/UploadPaperDto"}}}}, "responses": {"200": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SciPaper"}}}}}}}, "//": {"get": {"description": "", "operationId": "getAll", "parameters": [], "responses": {"200": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/SciPaper"}}}}}}}}, "//{id}": {"get": {"description": "", "operationId": "getOne", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}], "responses": {"200": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/SciPaper"}}}}}}}, "/publishpaper/{sciPaper}": {"get": {"description": "", "operationId": "publishPaper", "parameters": [], "requestBody": {"description": "JSON representation object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "responses": {"200": {"description": "", "content": {}}}}}}, "components": {"schemas": {"SciPaper": {"type": "object", "properties": {"id": {"type": "string"}, "author": {"type": "string"}, "title": {"type": "string"}, "sections": {"type": "array", "items": {"$ref": "#/components/schemas/Section"}}}, "required": ["id", "author", "title"]}, "Section": {"type": "object", "properties": {"name": {"type": "string"}, "content": {"type": "string"}}, "required": ["name", "content"]}, "UploadPaperDto": {"type": "object", "properties": {"title": {"type": "string"}, "sections": {"type": "array", "items": {"$ref": "#/components/schemas/Section"}}}, "required": ["title", "sections"]}}, "requestBodies": {"SciPaperBody": {"description": "JSON representation of SciPaper object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "SectionBody": {"description": "JSON representation of Section object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}, "UploadPaperDtoBody": {"description": "JSON representation of UploadPaperDto object", "required": true, "content": {"application/json": {"schema": {"$ref": "#components/schemas/UploadPaperDto"}}}}}, "responses": {"SciPaperOK": {"description": "JSON representation of SciPaper object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/SciPaper"}}}}, "SectionOK": {"description": "JSON representation of Section object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/Section"}}}}, "UploadPaperDtoOK": {"description": "JSON representation of UploadPaperDto object", "content": {"application/json": {"schema": {"$ref": "#components/schemas/UploadPaperDto"}}}}}}}