# silvera/sciLibrary/output/SciPaper/src/main/java/com/silvera/SciPaper/domain/model/SciPaper.java
/**
    THIS IS GENERATED CODE AND SHOULD NOT BE CHANGED MANUALLY!!!

    Generated by: silvera
    Date: 2022-05-07 18:49:21
*/

package com.silvera.SciPaper.domain.model;

import org.springframework.data.annotation.Id;
import javax.validation.constraints.*;
public class SciPaper {

    
    
    
    @Id
    
    @NotBlank(message="Field 'id' cannot be empty!")
    private java.lang.String id;
    
    
    @NotBlank(message="Field 'author' cannot be empty!")
    private java.lang.String author;
    
    
    @NotBlank(message="Field 'title' cannot be empty!")
    private java.lang.String title;
    
    private java.util.List<Section> sections= java.util.Collections.emptyList();


    
    public java.lang.String getId() {
        return this.id;
    }

    public void setId(java.lang.String id) {
        this.id = id;
    }
    
    public java.lang.String getAuthor() {
        return this.author;
    }

    public void setAuthor(java.lang.String author) {
        this.author = author;
    }
    
    public java.lang.String getTitle() {
        return this.title;
    }

    public void setTitle(java.lang.String title) {
        this.title = title;
    }
    
    public java.util.List<Section> getSections() {
        return this.sections;
    }

    public void setSections(java.util.List<Section> sections) {
        this.sections = sections;
    }
    

    

}